Optimize  config/loader
