+++ config/storage/file/file_test.go
@@ -1,0 +1,22 @@
+package file
+
+import (
+	"bytes"
+	"testing"
+)
+
+func TestFile_Write_Load(t *testing.T) {
+	data := []byte("123456")
+	storage := NewStorage("/tmp/test.file")
+	if err := storage.Write(data); err != nil {
+		t.Fatal(err)
+	}
+
+	d, err := storage.Load()
+	if err != nil {
+		t.Fatal(err)
+	}
+	if bytes.Compare(data, d) != 0 {
+		t.Fatal(data, d)
+	}
+}
